Fix issue where the library name was not being considered when outputting the templated matching header
I ran into an issue where the plugin for Templated Matching headers did not correctly output the import of the main object header. Instead of using the library name from the algebraic type, it simply used Maybe.Nothing(), which ended up outputting a local import and not a framework-style import.

Acceptance and unit tests pass.