Add nil checks for nonnull coding attributes
Remodel's generated `NSCoding` code, especially in combination with `RMAssumeNonnull`, is unsafe because it does not check for `nil` when initializing `nonnull` properties. This leads to instances that lie about their nullability, which is particularly bad when consuming a Remodel class from Swift.

To solve this, I've added `nil` checks in `-initWithCoder:`, and return `nil` from the initializer if any `nonnull` property is `nil`. The return type of `-initWithCoder:` is `nullable instancetype`, so this is reasonable and clients should be aware of a potential `nil` return value.