Add codingKey attribute to RMCoding
Summary:
The `RMCoding` plugin currently has a `legacyCodingKey` attribute that allows you to use arbitrary strings as fallback coding keys. However, all encoding is done to a key that Remodel decides on (`CAPITALIZED_SNAKE_CASE`). For people migrating to Remodel, there's not any need for this, and it could be harmful if they're trying to retain compability with older versions of their app (e.g. macOS document-based apps, shared containers with App Groups).

I've also created an `AttributeMap` type to represent the nesting of `Attribute` values inside other types, as I was introducing a new usage, and there were a few repeated definitions of the type.

Reviewed By: calimarkus

Differential Revision: D10085309

fbshipit-source-id: 138f7c813f247fb7f1463f9b4fdc23d529fc922b