Cast to id after decoding in RMCoding to silence nullability errors
Summary:
The `RMCoding` plugin assigns the result of decoding method directly to ivars. However, these methods return `nullable` for objects. If `-Wnullable-to-nonnull-conversion` is enabled, the compiler fails (assuming `-Werror`) on these files. We want to be able to have that flag on for a library, while still allowing the use of Remodel.

I considered just using the `.valueObjectConfig` option to disable compiler flags, but decided against it because localizing this to the `RMCoding` plugin allows us to retain that check for our other generated code, giving us additional safety benefits in the code generator.

The long-term fix would be an `NSCoding` plugin that performed validation, but there are backwards-compatibility issues, so it would probably need to be a separate plugin.

Differential Revision: D10367360

fbshipit-source-id: 7dc3ceb2d86a3b1a01b3afbef04fa5daa490def0