Add plugin to generate ivars instead of properties
Summary:
Using properties causes the compiler to emit property metadata, increasing the size of apps using Remodel. This diff adds an alternative to the standard `RMImmutableProperties` that generates an ivar and a "manual" getter instead, removing the property metadata: `RMImmutableIvars`. In addition to emitting the ivars and the getters, the plugin is also responsible for emitting the initializer, which it does from a shared function that was extracted in the previous diff. The interface that is exposed to typical callers is effectively unchanged: they can continue to access value object attributes with dot notation as if they were still properties.

This also required making value object plugins be capable of emitting ivars, previously only ADT plugins could. This was an easy change to make, as the underlying renderer was just hard-coded to return `[]`, I changed it to call an optional function.

Reviewed By: ColinCampbell

Differential Revision: D14146191

fbshipit-source-id: f96b51556cd0c1391cf183c23a47e7170a27d65e