Don't use inline flag for block types. Instead, put it in the class if that's where it belongs.
Summary: We were previously marking block types as 'inline'. What this really meant is put it inside the class definition. But with single-file emission, it was inserting it into EVERY class in the file, instead of the one class it belonged with. Now the plugin will put it right into the class and we output it properly.

Reviewed By: calimarkus

Differential Revision: D14467046

fbshipit-source-id: a35297df2e45f95f909b0ec44529be8931746386