Enable strict null checks
Summary:
Typescript has an option for strict null checks, but we haven't had it enabled. I enabled it in the build script, then went through and fixed all of the errors, either by actually fixing them, or by suppressing them with `!`.

This means that `const foo: string = null` is no longer valid, as well as variants with function parameters etc.

Reviewed By: ColinCampbell

Differential Revision: D17327485

fbshipit-source-id: fd1d82b9aa128b955c8214f303849a3da122ad3c