Use __unsafe_unretained for local blocks in generic matchers
Summary:
This will reduce overall binary size and improve efficiency a tiny bit, as ARC will no longer unnecessarily copy these blocks to the heap. They're only used as local variables within the method and passed to the match functions, which run synchronously, so this is safe to do.

Also updated to template strings instead of concatenation.

Reviewed By: FieryRobot

Differential Revision: D18080038

fbshipit-source-id: 160043c4f9c634815617da0a569a90fee14c345d