Remove weird special case for self-references
Summary:
We specifically add a forward declaration for the case where an object has an attribute of its own type. This is perfectly legal Objective-C:

```
interface Foo : NSObject
property (nonatomic) Foo *foo; // look ma, no forward declaration
end
```

Turns out this *is* needed specifically for the case of an ADT that is self-referential, since we use `typedef` to declare a matcher block type before the algebraic data type is declared.

Move it into `algebraic-type-matching-generic`.

Reviewed By: ColinCampbell

Differential Revision: D20804880

fbshipit-source-id: 3b8bc5579140f8d4e1a569fa663af2e886a9a7b5