Tweak parser for empty generic sections
Summary:
Types like the following occasionally crop up:

```
%type Foo canForwardDeclare=false
...
Foo<> foo;
```

where `Foo` is actually a C++ class:

```
template<typename... T>
class Foo {...}
```

This is sort of an abuse of Remodel. Oh well.

This is currently parsed as `conformingProtocol=""` which doesn't make sense. Parse it as an empty generics section instead.

Reviewed By: natestedman

Differential Revision: D20825684

fbshipit-source-id: 8b4baab96a7bffaaf06771aae417961167d95947