Respect the adtConfigPath path
Summary: We aren't respecting this parameter at all. This diff fixes that. Rather than furthering the divide between two files that are similar but have a lot of forked logic (object-specs.ts and algebraic-types.ts), I tried to push the complexity into shared code (findConfig), but simply letting the caller decide to disallow searching for file, and it must be present in the directory.

Reviewed By: natestedman

Differential Revision: D22777174

fbshipit-source-id: 5e89244b02c2504e4f5e28d193117869270f973f