Use `failWithError:` instead of raising an exception
Summary:
`NSKeyedUnarchiver` has two policies defined which indicate how decoding failures should be handled:

- `NSDecodingFailurePolicyRaiseException`, which throws an `NSException` when there is a decoding problem (e.g. decoding an object as an integer)
- `NSDecodingFailurePolicySetErrorAndReturn`, which returns from decoding with an `NSError` provided to the top-level caller

`NSException` throwing has been the historical default, but newer API on `NSKeyedUnarchiver` now opt into the `NSError` handling policy by default.

When implementing `initWithCoder:`, if an `NSException` is thrown, it bypasses the `NSCoder` policy and may cause an app to crash if it's not handled. This happens even if the current decoding policy is `NSDecodingFailurePolicySetErrorAndReturn`.

Thankfully, whichever policy is active there is a single API on `NSCoder` which handles either case: `failWithError:`. Passing an `NSError` to this API (and returning `nil`) will cause an `NSException` to be raised, or will cause the `NSError` to be propagated, depending on the current policy.

In Remodel, we only flag problems with decoding algebraic types when their encoded subtype is not recognized. This can happen if a new version of an app has removed a subtype. So, let's use this API on `NSCoder` instead of forcibly raising an exception.

The format of the `NSError` is modeled after similar errors that happen when there's a type mismatch (e.g. decoding an object as an integer):

```
Error Domain=NSCocoaErrorDomain
  Code=NSCoderReadCorruptError
  UserInfo={NSDebugDescription=*** -[NSKeyedUnarchiver decodeInt64ForKey:]: value for key (KEYNAME) is not an integer number}
```

Reviewed By: natestedman

Differential Revision: D26237572

fbshipit-source-id: 9104f2596306c76d887df954f8326cdf96a3e8af