Add method to discern types that need annotations.
Summary:
This diff adds a method to check if a type should have a nullability annotation. Right now, Remodel relies on an "inherited" value in the NullabilityType enum, and that enum is used to mean different things in different places.

Submitting this for review separately from the rest of the stack, since it's the core logic behind the rest of the fixes in this stack.

Reviewed By: natestedman

Differential Revision: D33006005

fbshipit-source-id: d3e4269fa57ffcd6f5ef38beb43e9587dc460e54